requirements.txt

streamlit
streamlit-drawable-canvas
google-generativeai
Pillow
pandas
plotly
